bootspec: schema_version is u32
JSON can only display (2^53)-1 precisely, so limit the number of
possible versions to (2^32)-1 (which should be more than enough).

+++ bootspec/src/lib.rs
@@ -20,6 +20,6 @@ pub struct BootJsonPath(pub PathBuf);
 /// The current bootspec schema.
 pub type BootJson = v1::BootJsonV1;
 /// The current bootspec schema version.
-pub const SCHEMA_VERSION: usize = 1;
+pub const SCHEMA_VERSION: u32 = 1;
 /// The current bootspec schema filename.
 pub const JSON_FILENAME: &str = "boot.v1.json";
